+++ CHANGELOG.md
@@ -1,4 +1,36 @@
 <!-- markdownlint-disable MD041 -->
+## 2.0.0 (Unrelease)
+
+BREAKING CHANGES:
+
+* dependencies: updating to `v2.0.0` minimum of terraform.
+* dependencies: updating to `v3.40.0` minimum of hashicorp/azurerm provider.
+
+ENHANCEMENTS:
+
+* The `sku.capacity` parameter changes the `string` type to `number` and is now an optional parameter.
+* The `autoscale_configuration.min_capacity` parameter changes type `string` to `number`.
+* The `autoscale_configuration.max_capacity` parameter changes type `string` to `number`.
+* The `sku.capacity` parameter is now a single parameter with the name `capacity`.
+
+FEATURES:
+
+* **New Parameter**: `firewall_policy_id`
+* **New Parameter**: `capacity`
+* **New Parameter**: `zones`
+* **New Parameter**: `sku_name`
+* **New Parameter**: `enable_http2`
+
+BUG FIXES:
+
+DEPRECATIONS:
+
+[waf-config-deprecate]: https://azure.microsoft.com/en-us/updates/retirement-support-for-application-gateway-web-application-firewall-v2-configuration-is-ending/
+[appgw-sku-deprecate]: https://azure.microsoft.com/en-us/updates/application-gateway-v1-will-be-retired-on-28-april-2026-transition-to-application-gateway-v2/
+
+* **Parameter**: `waf_configuration` is deprecated in favor of `firewall_policy_id`. WAF configuration must now be performed using a firewall policy. [Retirement: Support for Application Gateway Web Application Firewall v2 Configuration is ending][waf-config-deprecate].
+* **Parameter**: `sku` is deprecated in favor of `sku_name`. The `Standard_Small`, `Standard_Medium`, `Standard_Large` and `WAF_Medium` sku types are also deprecated. [Application Gateway V1 will be retired on 28 April 2026– Transition to Application Gateway V2][appgw-sku-deprecate].
+
 ## 1.2.0 (January 27, 2022)
 
 FEATURES: